aarch32: add virtual timer control support
Define virtual timer resources for a platform to be able to control
virtual timers, even if TF-A does not enable them.
This is required for example when the secure world wants to reset
the system. Virtual timers enabled from non-secure world can
jeopardize the reset sequence due to pending interrupts.

Change-Id: Ieb0ce4819809ae9c4ab11f9244a6cacf9437d6b4
Signed-off-by: Nicolas Le Bayon <nicolas.le.bayon@st.com>
Signed-off-by: Yann Gautier <yann.gautier@st.com>